Estimation of volume of the food item present in the 2-D images usingvarious methods.
World-wide obesity has been tripled since 1975 and
roughly 45% of all American are obese while 70% of all
adults in America are over-weight. 60% of all population
and 95% of Americans have access to the internet. Hence,
if the calorific value of each meal can be easily calculated,
especially from a single picture captured in a smart phone,
it could be help those who want to make healthier life decisions.
There are various techniques which can be used
for calculating the volume of food found in a pictures. A
few of these methods like using birds-eye view, frame transforms,
structure-from-motion, neural radiance field etc. are
compared.